Make readData method to AdaptHiveBaseArcticDataReader to let other subclass reader use it. Change Trino's filter to AdaptHiveArcticDeleteFilter.
