+++ trino/src/main/java/com/netease/arctic/trino/keyed/KeyedConnectorPageSource.java
@@ -21,6 +21,7 @@ package com.netease.arctic.trino.keyed;
 import com.google.common.collect.ImmutableList;
 import com.netease.arctic.data.DataFileType;
 import com.netease.arctic.data.PrimaryKeyedFile;
+import com.netease.arctic.hive.io.reader.AdaptHiveArcticDeleteFilter;
 import com.netease.arctic.io.reader.ArcticDeleteFilter;
 import com.netease.arctic.scan.ArcticFileScanTask;
 import com.netease.arctic.table.MetadataColumns;
@@ -74,7 +75,7 @@ public class KeyedConnectorPageSource implements ConnectorPageSource {
   private DynamicFilter dynamicFilter;
   private TypeManager typeManager;
   private FileIoProvider fileIoProvider;
-  private ArcticDeleteFilter<TrinoRow> arcticDeleteFilter;
+  private AdaptHiveArcticDeleteFilter<TrinoRow> arcticDeleteFilter;
 
   private List<ColumnHandle> requireColumnsDummy;
   private Type[] requireColumnTypes;
@@ -97,7 +98,7 @@ public class KeyedConnectorPageSource implements ConnectorPageSource {
       DynamicFilter dynamicFilter,
       TypeManager typeManager,
       FileIoProvider fileIoProvider,
-      ArcticDeleteFilter<TrinoRow> arcticDeleteFilter) {
+      AdaptHiveArcticDeleteFilter<TrinoRow> arcticDeleteFilter) {
     this.expectedColumns = expectedColumns;
     this.icebergPageSourceProvider = icebergPageSourceProvider;
     this.transaction = transaction;
