+++ trino/src/main/java/com/netease/arctic/trino/keyed/KeyedPageSourceProvider.java
@@ -21,6 +21,7 @@ package com.netease.arctic.trino.keyed;
 import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
 import com.netease.arctic.data.PrimaryKeyedFile;
+import com.netease.arctic.hive.io.reader.AdaptHiveArcticDeleteFilter;
 import com.netease.arctic.io.reader.ArcticDeleteFilter;
 import com.netease.arctic.scan.ArcticFileScanTask;
 import com.netease.arctic.scan.KeyedTableScanTask;
@@ -93,7 +94,7 @@ public class KeyedPageSourceProvider implements ConnectorPageSourceProvider {
         .filter(column -> !columns.contains(column))
         .forEach(requiredColumnsBuilder::add);
     List<IcebergColumnHandle> requiredColumns = requiredColumnsBuilder.build();
-    ArcticDeleteFilter<TrinoRow> arcticDeleteFilter = new KeyedDeleteFilter(
+    AdaptHiveArcticDeleteFilter<TrinoRow> arcticDeleteFilter = new KeyedDeleteFilter(
         keyedTableScanTask,
         tableSchema,
         requiredColumns,
